+++ alembic/versions/f9a2f1c1d378_create_subscriber_table.py
@@ -9,7 +9,6 @@ Create Date: 2024-04-04 16:47:49.974867
 from typing import Sequence, Union
 
 import sqlalchemy as sa
-from sqlalchemy.dialects.postgresql import UUID
 
 from alembic import op
 
@@ -23,9 +22,8 @@ depends_on: Union[str, Sequence[str], None] = None
 def upgrade() -> None:
     op.create_table(
         "subscriber",
-        sa.Column("id", UUID(as_uuid=True), primary_key=True),
-        sa.Column("subscriber_type", sa.String(50), nullable=False),
-        sa.Column("webhook_url", sa.String(50), nullable=False),
+        sa.Column("subscriber_type", sa.String(50), nullable=False, primary_key=True),
+        sa.Column("webhook_url", sa.String(50), nullable=False, primary_key=True),
     )
