+++ laa_crime_application_store_app/models/subscriber_schema.py
@@ -1,6 +1,4 @@
-import uuid
-
-from sqlalchemy import UUID, Column, String
+from sqlalchemy import Column, String
 
 from laa_crime_application_store_app.data.database import Base
 
@@ -8,6 +6,9 @@ from laa_crime_application_store_app.data.database import Base
 class Subscriber(Base):
     __tablename__ = "subscriber"
 
-    id = Column(UUID, primary_key=True, default=uuid.uuid4)
-    subscriber_type = Column(String, nullable=False)
-    webhook_url = Column(String, nullable=False)
+    subscriber_type = Column(String, nullable=False, primary_key=True)
+    webhook_url = Column(
+        String,
+        nullable=False,
+        primary_key=True,
+    )
