+++ laa_crime_application_store_app/routers/v1/application.py
@@ -13,7 +13,6 @@ from laa_crime_application_store_app.schema.application import Application as Ap
 from laa_crime_application_store_app.schema.application_new import ApplicationNew
 from laa_crime_application_store_app.schema.application_update import ApplicationUpdate
 from laa_crime_application_store_app.schema.basic_application import ApplicationResponse
-from laa_crime_application_store_app.schema.subscriber import Subscriber
 from laa_crime_application_store_app.services.v1.application_service import (
     ApplicationService,
 )
@@ -109,43 +108,3 @@ async def put_application(
     logger.info("APPLICATION_UPDATED", application_id=application.application_id)
     NotificationService().notify(db, request, application.application_id)
     return Response(status_code=201)
-
-
-@router.post("/subscriber")
-@auth_logger
-async def post_subscriber(
-    request: Request,
-    subscriber: Subscriber,
-    db: Session = Depends(get_db),
-):
-    logger.info("CREATING_SUBSCRIBER")
-    new_subscriber = NotificationService.subscribe(
-        db, subscriber.subscriber_type, subscriber.webhook_url
-    )
-
-    if new_subscriber is not None:
-        logger.info("SUBSCRIBER_CREATED")
-        return Response(status_code=201)
-    else:
-        logger.info("SUBSCRIBER_ALREADY_EXISTED")
-        return Response(status_code=204)
-
-
-@router.delete("/subscriber")
-@auth_logger
-async def delete_subscriber(
-    request: Request,
-    subscriber: Subscriber,
-    db: Session = Depends(get_db),
-):
-    logger.info("DELETING_SUBSCRIBER")
-    deleted = NotificationService.unsubscribe(
-        db, subscriber.subscriber_type, subscriber.webhook_url
-    )
-
-    if deleted:
-        logger.info("SUBSCRIBER_DELETED")
-        return Response(status_code=204)
-    else:
-        logger.info("SUBSCRIBER_NOT_FOUND")
-        return Response(status_code=404)
