+++ tests/routers/v1/application_test.py
@@ -10,7 +10,6 @@ from laa_crime_application_store_app.models.application_schema import Applicatio
 from laa_crime_application_store_app.models.application_version_schema import (
     ApplicationVersion,
 )
-from laa_crime_application_store_app.models.subscriber_schema import Subscriber
 
 logger = structlog.getLogger(__name__)
 
@@ -129,6 +128,7 @@ def test_post_application_returns_duplicate_error_if_id_already_exists(
 def test_post_application_notifies_subscribers(
     client: TestClient, dbsession: Session, seed_subscriber, httpx_mock
 ):
+    # This will throw an AssertionError if the url specified has not been called by the time the test ends
     httpx_mock.add_response(url=seed_subscriber.webhook_url)
 
     client.post(
@@ -405,6 +405,7 @@ def test_put_application_notifies_subscribers(
     seed_subscriber,
     httpx_mock,
 ):
+    # This will throw an AssertionError if the url specified has not been called by the time the test ends
     httpx_mock.add_response(url=seed_subscriber.webhook_url)
 
     client.put(
@@ -419,62 +420,3 @@ def test_put_application_notifies_subscribers(
             "application": {"id": 10},
         },
     )
-
-
-def test_post_subscriber_creates_new_subscriber(client: TestClient, dbsession: Session):
-    response = client.post(
-        "/v1/subscriber",
-        headers={"Content-Type": "application/json"},
-        json={
-            "webhook_url": "https://example.com/webhook",
-            "subscriber_type": "provider",
-        },
-    )
-    subscriber = dbsession.query(Subscriber).first()
-    assert subscriber.webhook_url == "https://example.com/webhook"
-    assert subscriber.subscriber_type == "provider"
-    assert response.status_code == 201
-
-
-def test_post_subscriber_avoids_duplicates(
-    client: TestClient, dbsession: Session, seed_subscriber
-):
-    response = client.post(
-        "/v1/subscriber",
-        headers={"Content-Type": "application/json"},
-        json={
-            "webhook_url": seed_subscriber.webhook_url,
-            "subscriber_type": seed_subscriber.subscriber_type,
-        },
-    )
-    assert dbsession.query(Subscriber).count() == 1
-    assert response.status_code == 204
-
-
-def test_delete_subscriber_removes_record(
-    client: TestClient, dbsession: Session, seed_subscriber
-):
-    response = client.request(
-        "DELETE",
-        "/v1/subscriber",
-        headers={"Content-Type": "application/json"},
-        json={
-            "webhook_url": seed_subscriber.webhook_url,
-            "subscriber_type": seed_subscriber.subscriber_type,
-        },
-    )
-    assert dbsession.query(Subscriber).count() == 0
-    assert response.status_code == 204
-
-
-def test_delete_subscriber_fails_gracefully(client: TestClient, dbsession: Session):
-    response = client.request(
-        "DELETE",
-        "/v1/subscriber",
-        headers={"Content-Type": "application/json"},
-        json={
-            "webhook_url": "https://example.com/webhook",
-            "subscriber_type": "provider",
-        },
-    )
-    assert response.status_code == 404
